requirements.txt
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
streamlit>=1.28.0
tqdm>=4.66.0
